refactor: 삭제시 검증 로직 추가 #8

+++ be/src/main/java/buddyguard/mybuddyguard/weight/service/WeightService.java
@@ -63,10 +63,12 @@ public class WeightService {
     }
 
     @Transactional
-    public void deleteWeightRecord(Long id) {
+    public void deleteWeightRecord(Long id, Long petId) {
 
         Weight weight = weightRepository.findById(id)
                 .orElseThrow(RecordNotFoundException::new);
+
+        weight.validateOwnership(petId);
 
         weightRepository.delete(weight);
     }
